pyproject.toml
[tool.basedpyright]
typeCheckingMode = "standard"       # standard Python3, with everything else set to optional

[tool.yapf]
based_on_style = "pep8"
column_limit = 99

[tool.black]
line-length = 99
skip-string-normalization = true

[tool.poetry]
name = "exercises"
version = "0.1.0-alpha.1"
description = "StrongerFox exercise db"
authors = ["thodnev <thodnev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
ruamel-yaml = "^0.18.6"
rapidfuzz = "^3.11.0"
datamodel-code-generator = "^0.26.4"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "poetry"
major_version_zero = true

